requirements.txt
ultralytics
fastapi
uvicorn[standard]
pydantic
python-multipart
opencv-python-headless
torch
Pillow
